backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
openai==0.28.1
Pillow==10.0.1
pytesseract==0.3.10
pdf2image==1.16.3
python-dotenv==1.0.0
Werkzeug==2.3.7 